added build_binary project
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/branches/hieu_win32@4371 1f5c12ca-751b-0410-a591-d2e778427230

+++ kenlm/lm/build_binary.cc
@@ -87,7 +87,8 @@ void ShowSizes(const char *file, const lm::ngram::Config &config) {
     prefix = 'G';
     divide = 1 << 30;
   }
-  long int length = std::max<long int>(2, lrint(ceil(log10(max_length / divide))));
+
+  long int length = std::max<long int>(2, lrint(ceil(log10((float)max_length / (float)divide))));
   std::cout << "Memory estimate:\ntype    ";
   // right align bytes.  
   for (long int i = 0; i < length - 2; ++i) std::cout << ' ';
